Actualiza tipos de datos en el modelo de Producto y Venta; agrega la página de Terminal de Venta con funcionalidad de búsqueda y gestión de carrito.

+++ Data/Producto.cs
@@ -27,7 +27,7 @@ public class Producto
 
     [Required]
     [Column("stock_actual")]
-    public int StockActual { get; set; }
+    public decimal StockActual { get; set; }
 
     [Column("categoria_id")]
     public long? CategoriaId { get; set; }
